requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
pdfplumber
sentence-transformers
regex
pydantic
streamlit
requests